+++ src/main/java/znu/visum/components/genres/usecases/getpage/application/GetPageGenreController.java
@@ -3,17 +3,13 @@ package znu.visum.components.genres.usecases.getpage.application;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
-import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.web.SortDefault;
 import org.springframework.hateoas.server.ExposesResourceFor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
-import znu.visum.components.genres.domain.models.Genre;
 import znu.visum.components.genres.usecases.getpage.domain.GetPageGenreService;
 import znu.visum.core.pagination.application.GetPageResponse;
-import znu.visum.core.pagination.infrastructure.PageSearch;
-import znu.visum.core.pagination.infrastructure.SearchSpecification;
 
 @RestController
 @RequestMapping(value = "/api/genres", produces = MediaType.APPLICATION_JSON_VALUE)
@@ -32,19 +28,9 @@ public class GetPageGenreController {
   public GetPageResponse<GenreFromPageResponse> getPage(
       @RequestParam(required = false, defaultValue = "0") int offset,
       @RequestParam(required = false, defaultValue = "20") int limit,
-      @RequestParam(required = false, defaultValue = "") String search,
+      @RequestParam(required = false, defaultValue = "type=%%") String search,
       @SortDefault Sort sort) {
-    Specification<Genre> searchSpecification = SearchSpecification.parse(search);
-
-    PageSearch<Genre> pageSearch =
-        new PageSearch.Builder<Genre>()
-            .search(searchSpecification)
-            .offset(offset)
-            .limit(limit)
-            .sorting(sort)
-            .build();
-
     return GetPageResponse.from(
-        getPageGenreService.findPage(pageSearch), GenreFromPageResponse::from);
+        getPageGenreService.findPage(limit, offset, sort, search), GenreFromPageResponse::from);
   }
 }