+++ src/main/java/znu/visum/components/genres/usecases/getpage/domain/GetPageGenreService.java
@@ -1,9 +1,9 @@
 package znu.visum.components.genres.usecases.getpage.domain;
 
+import org.springframework.data.domain.Sort;
 import znu.visum.components.genres.domain.models.Genre;
 import znu.visum.core.pagination.domain.VisumPage;
-import znu.visum.core.pagination.infrastructure.PageSearch;
 
 public interface GetPageGenreService {
-  VisumPage<Genre> findPage(PageSearch<Genre> page);
+  VisumPage<Genre> findPage(int limit, int offset, Sort sort, String search);
 }