+++ src/main/java/znu/visum/components/movies/usecases/getpage/domain/GetPageMovieService.java
@@ -1,9 +1,10 @@
 package znu.visum.components.movies.usecases.getpage.domain;
 
+import org.springframework.data.domain.Sort;
 import znu.visum.components.movies.domain.models.Movie;
 import znu.visum.core.pagination.domain.VisumPage;
-import znu.visum.core.pagination.infrastructure.PageSearch;
 
 public interface GetPageMovieService {
-  VisumPage<Movie> findPage(PageSearch<Movie> page);
+  // TODO Spring Sort is not supposed to be here!
+  VisumPage<Movie> findPage(int limit, int offset, Sort sort, String search);
 }