+++ src/main/java/znu/visum/components/movies/usecases/getpage/domain/GetPageMovieServiceImpl.java
@@ -1,11 +1,11 @@
 package znu.visum.components.movies.usecases.getpage.domain;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import znu.visum.components.movies.domain.models.Movie;
 import znu.visum.components.movies.domain.ports.MovieRepository;
 import znu.visum.core.pagination.domain.VisumPage;
-import znu.visum.core.pagination.infrastructure.PageSearch;
 
 @Service
 public class GetPageMovieServiceImpl implements GetPageMovieService {
@@ -17,7 +17,7 @@ public class GetPageMovieServiceImpl implements GetPageMovieService {
   }
 
   @Override
-  public VisumPage<Movie> findPage(PageSearch<Movie> pageSearch) {
-    return movieRepository.findPage(pageSearch);
+  public VisumPage<Movie> findPage(int limit, int offset, Sort sort, String search) {
+    return movieRepository.findPage(limit, offset, sort, search);
   }
 }