+++ src/main/java/znu/visum/components/people/actors/domain/ports/ActorRepository.java
@@ -1,13 +1,13 @@
 package znu.visum.components.people.actors.domain.ports;
 
+import org.springframework.data.domain.Sort;
 import znu.visum.components.people.actors.domain.models.Actor;
 import znu.visum.core.pagination.domain.VisumPage;
-import znu.visum.core.pagination.infrastructure.PageSearch;
 
 import java.util.Optional;
 
 public interface ActorRepository {
-  VisumPage<Actor> findPage(PageSearch<Actor> page);
+  VisumPage<Actor> findPage(int limit, int offset, Sort sort, String search);
 
   Optional<Actor> findById(long id);