+++ src/main/java/znu/visum/components/people/directors/domain/ports/DirectorRepository.java
@@ -1,13 +1,13 @@
 package znu.visum.components.people.directors.domain.ports;
 
+import org.springframework.data.domain.Sort;
 import znu.visum.components.people.directors.domain.models.Director;
 import znu.visum.core.pagination.domain.VisumPage;
-import znu.visum.core.pagination.infrastructure.PageSearch;
 
 import java.util.Optional;
 
 public interface DirectorRepository {
-  VisumPage<Director> findPage(PageSearch<Director> page);
+  VisumPage<Director> findPage(int limit, int offset, Sort sort, String search);
 
   Optional<Director> findById(long id);