+++ src/main/java/znu/visum/components/people/directors/usecases/getpage/domain/GetPageDirectorService.java
@@ -1,9 +1,9 @@
 package znu.visum.components.people.directors.usecases.getpage.domain;
 
+import org.springframework.data.domain.Sort;
 import znu.visum.components.people.directors.domain.models.Director;
 import znu.visum.core.pagination.domain.VisumPage;
-import znu.visum.core.pagination.infrastructure.PageSearch;
 
 public interface GetPageDirectorService {
-  VisumPage<Director> findPage(PageSearch<Director> page);
+  VisumPage<Director> findPage(int limit, int offset, Sort sort, String search);
 }