+++ src/main/java/znu/visum/components/reviews/usecases/getpage/domain/GetPageMovieReviewService.java
@@ -1,9 +1,9 @@
 package znu.visum.components.reviews.usecases.getpage.domain;
 
+import org.springframework.data.domain.Sort;
 import znu.visum.components.reviews.domain.models.Review;
 import znu.visum.core.pagination.domain.VisumPage;
-import znu.visum.core.pagination.infrastructure.PageSearch;
 
 public interface GetPageMovieReviewService {
-  VisumPage<Review> findPage(PageSearch<Review> page);
+  VisumPage<Review> findPage(int limit, int offset, Sort sort, String search);
 }