+++ src/main/java/znu/visum/components/reviews/usecases/getpage/domain/GetPageMovieReviewServiceImpl.java
@@ -1,11 +1,11 @@
 package znu.visum.components.reviews.usecases.getpage.domain;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import znu.visum.components.reviews.domain.models.Review;
 import znu.visum.components.reviews.domain.ports.ReviewRepository;
 import znu.visum.core.pagination.domain.VisumPage;
-import znu.visum.core.pagination.infrastructure.PageSearch;
 
 @Service
 public class GetPageMovieReviewServiceImpl implements GetPageMovieReviewService {
@@ -17,7 +17,7 @@ public class GetPageMovieReviewServiceImpl implements GetPageMovieReviewService 
   }
 
   @Override
-  public VisumPage<Review> findPage(PageSearch<Review> pageSearch) {
-    return reviewRepository.findPage(pageSearch);
+  public VisumPage<Review> findPage(int limit, int offset, Sort sort, String search) {
+    return reviewRepository.findPage(limit, offset, sort, search);
   }
 }