+++ src/test/java/znu/visum/components/actors/usecases/getpage/GetPageActorRouteIntegrationTest.java
@@ -43,24 +43,20 @@ public class GetPageActorRouteIntegrationTest {
     mvc.perform(
             get(
                     "/api/actors?sort=type&search=type={type}&limit={limit}&offset={offset}",
-                    "%25%25",
+                    "%%",
                     20,
                     0)
                 .contentType(MediaType.APPLICATION_JSON_VALUE))
         .andExpect(status().isForbidden());
   }
 
-  // TODO test more cases when the refactor is done on pagination
-
   @Test
   @WithMockUser
-  @Sql("/sql/insert_multiple_actors.sql")
+  @Sql(scripts = {"/sql/truncate_all_tables.sql", "/sql/insert_multiple_actors.sql"})
   @DisplayName(
-      "When only an empty search is provided, it should use default value: limit 20, offset 0, ascending sort on type")
+      "when only empty parameters are passed, it should use default value (limit 20, offset 0, ascending sort on type, search empty like on forename and name)")
   public void defaultCase_itShouldReturnA200Response() throws Exception {
-    mvc.perform(
-            get("/api/actors?search=type={type}", "%25%25")
-                .contentType(MediaType.APPLICATION_JSON_VALUE))
+    mvc.perform(get("/api/actors").contentType(MediaType.APPLICATION_JSON_VALUE))
         .andExpect(status().isOk())
         .andExpect(
             MockMvcResultMatchers.content()
@@ -76,6 +72,56 @@ public class GetPageActorRouteIntegrationTest {
                         + "{'id':5,'name': 'Mitchell', 'forename':'Radha'},"
                         + "{'id':6,'name': 'Bean', 'forename':'Sean'},"
                         + "{'id':7,'name': 'Robbins', 'forename':'Tim'}],"
+                        + "'totalPages':1,"
+                        + "'first':true,"
+                        + "'last':true}"));
+  }
+
+  @Test
+  @WithMockUser
+  @Sql(scripts = {"/sql/truncate_all_tables.sql", "/sql/insert_multiple_actors.sql"})
+  @DisplayName(
+      "when a forename and name is provided, it should return the actors with the forename or name")
+  public void givenAForenameAndName_itShouldReturnA200Response() throws Exception {
+    mvc.perform(
+            get("/api/actors?search=forename={forename},name={name}", "Robert", "De Niro")
+                .contentType(MediaType.APPLICATION_JSON_VALUE))
+        .andExpect(status().isOk())
+        .andExpect(
+            MockMvcResultMatchers.content()
+                .json(
+                    "{'current':0,"
+                        + "'size':20,"
+                        + "'totalElements':1,"
+                        + "'content':["
+                        + "{'id':4,'name': 'De Niro', 'forename':'Robert'}"
+                        + "],"
+                        + "'totalPages':1,"
+                        + "'first':true,"
+                        + "'last':true}"));
+  }
+
+  @Test
+  @WithMockUser
+  @Sql(scripts = {"/sql/truncate_all_tables.sql", "/sql/insert_multiple_actors.sql"})
+  @DisplayName(
+      "when a like forename or like name is provided (comma), it should return the actors containing like forename or like name")
+  public void givenALikeForenameOrLikeName_itShouldReturnA200Response() throws Exception {
+    mvc.perform(
+            get("/api/actors?search=forename={forename},name={name}", "%dha%", "%y%")
+                .contentType(MediaType.APPLICATION_JSON_VALUE))
+        .andExpect(status().isOk())
+        .andExpect(
+            MockMvcResultMatchers.content()
+                .json(
+                    "{'current':0,"
+                        + "'size':20,"
+                        + "'totalElements':3,"
+                        + "'content':["
+                        + "{'id':2,'name': 'Kyle', 'forename':'MacLachlan'},"
+                        + "{'id':3,'name': 'Hardy', 'forename':'Tom'},"
+                        + "{'id':5,'name': 'Mitchell', 'forename':'Radha'}"
+                        + "],"
                         + "'totalPages':1,"
                         + "'first':true,"
                         + "'last':true}"));